+++ src/array/codec/array_to_bytes/zfp/zfp_bitstream.rs
@@ -3,7 +3,7 @@ use std::ptr::NonNull;
 use zfp_sys::{bitstream, stream_close, stream_open};
 
 /// A `zfp` bitstream.
-pub struct ZfpBitstream(NonNull<bitstream>);
+pub(super) struct ZfpBitstream(NonNull<bitstream>);
 
 impl Drop for ZfpBitstream {
     fn drop(&mut self) {