+++ src/array/codec/array_to_bytes/zfp/zfp_stream.rs
@@ -9,7 +9,7 @@ use zfp_sys::{
 use super::ZfpMode;
 
 /// A `zfp` stream.
-pub struct ZfpStream(NonNull<zfp_stream>);
+pub(super) struct ZfpStream(NonNull<zfp_stream>);
 
 impl Drop for ZfpStream {
     fn drop(&mut self) {